backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
sqlalchemy>=1.4
pymysql>=1.0
pydantic>=1.10
pydantic-settings>=2.0.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3.0
alembic>=1.10
aiosmtplib>=1.1
email-validator>=1.3
python-dotenv>=1.0
python-multipart>=0.0.6
alembic==1.17.2
annotated-types==0.7.0
dnspython==2.8.0
email-validator==2.3.0
greenlet==3.2.4
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
numpy==2.3.5
pydantic==2.12.4
pydantic_core==2.41.5
SQLAlchemy==2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0